Setup up postgresql db container
Once again referring to this as a guide:
https://docs.docker.com/compose/rails/

It took a while to get it working. The examples use tmp/db for the
data, but this didn't work well for me - I think because the web
container "owns" and mounts the rails dir, so having the db
container own and mount rails/tmp/db inside there is problematic.

Get around it by puting the postgres data in ./.postgresql-data.