Show 'check your email' page after sign-up
It seems default devise behavious is just to send users to the login
page after sign-up. Since we're doing email confirmation of email
address, we should explain that to the user so they know what's
going on.

Not sure if it belongs in the home controller, but I guess it's good
enough for now.