Add controller for TiddlyHost Hub
Currently it just lists every site that is marked as searchable.

Decided on a hub layout where the results are boxes going across the
page instead of just down. It's a little unconventional, but I feel
it might work.

Also add will_paginate gem and use it for paginating the site lists.