Distinguish different types of 'site unavailable'
Instead of the same 404 message every time, be more sophisticated
about the response.

Blogspot.com handles non-existent blogs with a 404 and an offer to
create the blog, so I figure we can do something similar.

Also add some test coverage, and make it so the favicon and download
actions returns no body.