Don't serve a site if it doesn't look legit
It's not too hard to upload arbitrary content to Tiddlyhost. Let's
anticipate someone will try that. In this patch we add some checking
before serving a site to make sure it seems valid.

(We could also consider doing similar validations when uploading,
but it's not done here.)

It's not bullet-proof, since it wouldn't be hard to fool the
validator, but this might act as a deterrent for casual attempts.

Note the Nokogiri parser round tripping would probably make it hard
to serve non-html content, but I haven't tested that yet.

Also introduce an email support link and a new link for filing an
issue directly.

Re status code 418:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/418