Allow creation of TW Classic sites
It's initially hidden under an 'Advanced Options' click, so it's
clear that TiddlyWiki 5 is the usual best choice.

Adds a new table for empties. It seems like we should remember which
empty file a site was created with, so let's do it properly, and
introduce a model. It also makes the list of available empties more
robust.

We don't need the enabled field yet, but I've included it for
future use.

Also removes the default empty type from th_file since we don't need
it any more.