Replace dataTables in admin controller
DataTable is nice, but we have ~3000 Tiddlyspot sites now, and
fetching all of those at once for client side pagination and sorting
is not viable.

I started looking at hooking it up with ajax but it appeared quite
unpleasant and not easy, so I decided instead to dump it.

Includes:
- Remove dataTables packages from yarn
- Remove some js and scss includes
- Rework the admin controller to add sorting handling
- Refactor so the three list pages share a common code and views
- Some helpers and styles for pretty clickable column headings
- Create some helpers for a few common admin renderings
- Remove annoying doesnt_exist model method

(Still no test coverage for this stuff, but I'm okay with it.)

Issue #47.