Fix incorrect cert path in Makefile
It's a fixup for 3437df8, but I've started to wind back the reckless
history rewrites.

The impact was that `make cert` would needlessly create a new cert
every time it ran.

Also, I realized it's no longer needed when building, since the cert
doesn't go inside the container any more, so tweak the task deps.