Move shared site code into a concern
I debated how to handle the 'allow_nil' situation for the
delegations. The compromise is we have some allow_nil delegations
for Site records that aren't really needed in return for more shared
code. I think it's harmless enough - e.g. we can count on db
constraints to ensure user id is present for sites.