Don't use rails key file in docker build
Instead pass it in as a build arg.

Why? Even though it's not in the final container image, it would be
present in some of the layers, and hence could be extracted maybe,
which would be a problem when/if I make the container image public.

(Some day I'd like to remove the credentials.yml.enc from the image
as well, but will figure that out later.)

Also add a make task for dumping secrets which I used to help
recreate the encrypted secrets with a new key.