Use named param for shadow in tw_file
It's better for readability not to have mystery booleans in the
calling code.

Also introduce a method for choosing the store.

Note: I considered also removing all the shadow store support. It's
not for anything useful, especially given there's not a lot of
interest in Classic TWs so far, and it does add a lot of clutter.
Maybe in future I'll ditch it.