Add jsMath javascript library
There's an old plugin that uses this. Add it back so that Tiddlyspot
sites with that plugin will work again. (It fails un-gracefully so
these site were broken rather than just unable to render equations.)

I threw it in under rails/public, which seems to be good enough.

There was also a fonts-sprite directory, but I'm not sure where to
find that, nor how important it is.

Issue #123