Support tspot stubs & add a script to create them
If we know nothing about a tspot site except its name and that it is
expected to exist, that's a "stub" record.

"Unstubbing" a site means populating its details, such as its
password, whether it's public or private, and (as of recently)
adding its content to the main S3 store. Support doing that
automatically "just in time".

Also create a script that I'll use to create a record for all 100k
tiddlyspot site that ever existed.

Note: I'm trying to avoid "populating" all those sites, (though
there are some arguments for just doing it). The argument against
doing it is that a lot of them could be classified as dead at some
time in the not too distant future and finally put to rest.

Possible next steps include stopping checking the dreamhost bucket
at all for the existence of a TiddlySpot site, then deleting the
"phantom" non-existing sites and stop creating new ones.