Disregard blob timestamp in hub for unsaved tspots
Now we have blobs for all tiddlyspot sites, the blob timestamp is
irrelevant if the site was never saved.

This fixes incorrect hub sorting for 'sort by recent update' because
ancient tspot sites were considered to be recently updated because I
just created their blob.

(This was easier than I had thought it was going to be.)

Issue: #154