Remove "download empties" guide from README
Since we checked them in, there's no longer a need to describe how
to fetch them in the "quick start" guide in the readme.

Also because ansible was used just empty downloading, we don't need
to mention the need for ansible any more.

Also (unrelated fix while I was in there), explain that the local
dev environment doesn't send confirmation emails.