Use sudo for docker and docker-compose
I'm not sure why it changed, perhaps it used to use my bash alias
which defined docker-compose as `sudo docker-compose`, or maybe
something changed when I did a dnf upgrade. Now it seems to require
sudo docker-compose every time.

I'd like to switch to rootless docker (or podman) some time in the
future, but for now we'll stick with sudo.