Fix bootstrap icon colors
Not sure why, but the css 'color' stopped working for the icons, so
every icon was showing up as solid black.

Fix it by using opacity instead, or if that's no an option use a
fill attribute on the svg element.

It's a bit messy, but I want to fix it quickly.