Improved and better looking forms
Forms updated include:
- Sign in, sign up and the related devise forms
- Account edit
- Site create/edit
- Site upload
- Tiddlyspot claim
- Tiddlyspot site edit
- Tiddlyspot site change password

I'm not great with bootstrap, and this is still pretty scrappy, but
it's better than before.

Not sure about the layout partial that ended up being just a div,
but I'm leaving it for now. Originally I thought it needed
the form-group class, but that's actually gone in bootstrap 5.

Also includes numerous cosmetic fixups which I'm just going to
throw all together in one big commit, including:
- Use layout partials for more DRY form rendering
- Replace some custom css with standard bootstrap for margins and
  paddings. (More to come soon maybe).
- Some extra test coverage for the account related forms
- Use bg-light for the main body background color instead of white
- Improve the look of the footer, give it a darker bg
- Rearrange columns in sites index so actions is second
- Tweak the tiddlyspot home page text
- Remove unused site tags helper