Add ETag header
It's useful generally for caching IIUC, but also has some benefits
related to the put saver which I intend to make use of in upcoming
commits, particularly to use the If-Match header to avoid
unintentionally overwriting newer changes.

Not sure how useful it is for downloads and for the json and .tid end
points, but I guess it won't do any harm there.

Add it to tspot sites as well even though I don't plan to use the
put saver for those.