Do less work when building the prod image
If we precompile assets before building the prod image then we don't
need to run rails while building, which means we don't need the db
password or the master key, and that's a good thing because we don't
want to expose those if we can help it.

Required also to tweak an overzealous .dockerignore to ensure the
compiled assets get copied in to the image.