Rework secrets handling in the prod container
Remove the the master key value out of the image and out of the
docker history.

Using volumes to mount the secret files seems to be better and less
trouble than trying to wrangle environment variables while keeping
their values secret.

(Actually I had some trouble making environment variables work which
I realized later was largely due to the lack of sudo -E or
--preserve-env, but now I think this is preferrable regardless.)

I also realized that the way I was setting the master key
environment var wasn't actually working due to the sudo -E thing, a
problem which was masked due to the fact that the master key file
was actually still in the prod container image.

Also:
- Use the same technique for the encrypted credentials since we might
  as well remove them from the prod container image as well.
- A fix to make sure the prod container has a prerelease.html
  available when starting it locally.
- A changed location for the secrets dir when deployed up to prod
- Some ansible to deploy the credentials file so it can be mounted
  in the prod container along with the master key file.