Use a secure database password
I spent lots of time on more complicated versions of this, but then
I realized that the POSTGRES_PASSWORD var for the postgres container
is only used when there is no existing database, so there's no need
to get the secret into an environment var for the sake of the
postgre container.

Using environment vars might be more conventional here, but this
seems to be working fine and it means we have just one way to handle
secrets instead of two.