Record what kind a site is
There are two "kinds" currently, tw5, classic, but a new one will be
introduced in the next commit, which is why this is being done now.

(Since classic is 2.x and TW5 is 5.x, the versions don't clash,
which is why this wasn't so important before.)

The detection method is not very robust since it's using a regex
rather than the nokogiri doc, but should be good enough.