Thumbnail images for hub sites
Uses grover and puppeteer which were installed in the previous
commit. Update the thumbnail automatically on site save. Use a
background job to generate the screenshots since it's an expensive
operation.

Currently we're showing the thumbnail in the hub listing only. It
might appear in other places later but we'll start with this.

Includes some reworking of the hub view to make it look better with
the thumbnails.