Ansible tasks to customize docker /var/lib
The goal is so that we don't run out of space on the root partition
so easily. I'm sure there are other better ways to solve that
problem, but this will be an improvement for now.

Based on [1] but with some extra work to make sure the selinux
context is right.

[1] https://linuxconfig.org/how-to-move-docker-s-default-var-lib-docker-to-another-directory-on-ubuntu-debian-linux