Move node_modules for smaller image layers
The goal is that a small change in application code should produce a
small container layer change so we don't spent forever pushing and
pulling layers with 500MB of the same javascript.

It's a little less tidy that the way it's done for ruby gems since
yarn and webpacker are more particular about where they look for
node modules.