Grid view with thumbnails for sites index
There's now two different views, the existing "list" view and a new
"grid" view which shows thumbnails.

A cookie is used (rather than a user setting). to remember which
view you're using.

I'm trying to reuse/recycle css and partials from the hub views
rather than have two similar site tiles. As usual the html and css
are a bit scrappy.

There's also a change to the hub page - it's now using a flex box
with gap. I'm not 100% confident about that, but so far it seems fine.