Add action menu item to create a thumbnail
It's not essential, since updating a site will trigger a thumbnail
create, but this avoids needing to explain that to users.

Also fix a bug in the background job so doesn't assume everything is
a site, and hence works for Tiddlyspot sites.

Only show it in the new grid view so we don't confuse users looking
at the list view who maybe don't know about thumbnails at all.