Always use newest attachment when multiple exist
I'm observing that sometimes (a few times a month) when saving a
site the older tiddlywiki file attachment is not properly detached.
It's not clear how this happens and I haven't been able to reproduce
it.

The clean-attachment-dupes script removes the stale attachments. I
previously thought they were reasonably harmless, but I realized
recently that indeed they are harmful.

It depends on the sequence of events, but if someone saves a bunch
of changes, and the old attachment isn't removed properly, the
changes can be effectively lost.

The workaround: For a has_one_attachment there should only be one
record found, and rails use ..some arel expression..first to find
the attachment, and by default that returns the older attachment.

This patch attempts to make that `..first` method select the newer
attachment instead of the older one, which means everything should
work as expected when there are duplicate attachments.

I'll still want to clean up the duplicates, but if this works then
their existence becomes much less consequential.

It would be good to understand the root cause better, but at this
stage I'm just trying to reduce or hopefully eliminate the impact.

Issue: #245