Add a stable url for thumbnail images
The motivation is so it can be used with OpenGraph meta elements.

I guess it could be used for the thumbnail images in the hub and
grid view for consistency, but I decided not to do that. Maybe
there's some extra value (caching perhaps?) in using
rails_storage_proxy_path.