Prod makefile actions to clean dupe attachments
I'm still finding these a few times per month. Save a little time
cleaning them up.

(There's an issue open suggesting this could be done via the web UI,
and also it could be automated with a regular cleanup job, but I'm
hesitating because I'd really prefer to find the root cause and fix
that.)