Rework UI for choosing private/public/hub listed
Make it more clear that there are effectively three types of access
and add some explanations for each one.

Note that the change is UI only; the underlying db schema is the
same as before. There's some field wrangling with jQuery to handle
the translation between the UI and the real site attributes.

Issue #248