Add an action menu link to open in safe mode
Safe mode is a TiddlyWiki feature that disables plugins and certain
customizations on startup. This patch adds a link to the action menu
to open TiddlyWikis in safe mode.

Use target _blank for these links because safe mode won't work if
it's not opening in a fresh new tab, or at least that seems to be
the case.

Issue #268