Improve active user stat queries
The main fix was to use current_sign_in_at instead of
last_sign_in_at, which is actually the sign in before the most
recent sign in, so it was under-counting the numbers.

Also add scopes to the user model and tweak it a little to exclude
"logged in once" users, since I guess they're not particularly
active.