Viewing & interacting with a site's save history
The new history page lets you see a list of previous saved versions,
view and download them, restore them and discard them.

Implemented as new actions in the existing sites controller, but
they are defined in a concern to help keep things tidy.

The current plan is that this feature will be for subscribed users
only. (Currently it's hidden for everyone except admin user.)

Also:
- Make the site clickable in the mini tile used when cloning. That
  partial is recycled for use on the new history page, and I wanted
  it to be clickable there.

- Swap the order of the view and download menu items in the actions
  menu. It makes more sense that way on the history page, so change
  the main actions menu as well for consistency.

- Briefer option for nice timestamp helper