Avoid super long error output when debugging
Nokogiri doc inspect output for a TiddlyWiki file is extremely long,
and because of that I sometimes see error messages longer than the
scrollback history of my terminal. Try to avoid that.

Simple reproducer:
  >> Site.last.th_file.foo
  undefined method `foo' for #<ThFile:0x00007f72cc65e1b8
  @doc=#<Nokogiri::HTML4::Document:0x27d08 ...
  ...snip 1000s of screens of inspect output...

I considered monkey patching the class as per the following, but I
decided not to do that in case there were unexpected side-effects.

  Nokogiri::HTML4::Document.class_eval
    def inspect
      super.truncate(2500, separator: ',')
    end
  end

Note that String#truncate is a Rails method, not standard ruby.