Support external js version of TiddlyWiki
The core javascript is in a separate file so loading and saving is
quicker.

As well as faster save and load times, this should also reduce the
memory load on Tiddlyhost considerably since Tiddlyhost reads the
entire TiddlyWiki file into memory when serving a site.

Also tweak some nginx config so that the core js files get headers
to indicate they can be cached forever. Do the same for the rails
assets while we're in there, since they have unique digest-based
urls and they can and should be cached that way too.

Issue #171