Inject tiddlyhost prefix for external core urls
It's better if every site uses the same cached core js, instead of a
unique url per site.

Side-note: I spent a lot of time working on a client side version of
this where we define a custom macro value for coreURL. I eventually
got that to work, but then realized it could be done more simply
like this, making the TiddlyWiki customization unnecessary. There
might be some use for TiddlyWiki side coreURL customization later.

Issue: #171