New handy create button in header
I'm copying GitHub a little. I think it looks good, plus it saves
clicks when you're testing and hacking.

It doesn't look good on the small screen burger menu, so hide it
responsively with some bootstrap classes.

Also moved the admin link to the other side (again) and dimmed it.