Fix thumbnails for external core sites
The reason it didn't work is simply because the relative url to the
external core js file didn't work in the headless chromium. The fix
is a little hacky but should be good enough. I chose to do a string
regex instead of loading the full document (with nokogiri) and the
ThFile clss, since I want to keep the memory requirements down.

Issue #282