Promote the external core version of TiddlyWiki
Previously it was described as experimental, but I revising the
description to indicate it's ready for serious use. Hopefully not
prematurely.

The new explanations are aiming to be brief and not too technical,
but I think there should be some FAQ content to explain in more
detail later.

I'm being a little cautious and not making the default just yet, but
it is tempting. Soon maybe.