Fix hub query to handle multiple blobs per site
There's some interesting SQL here, but the net result hopefully is
that the hub query and the user site list query shows one row per
site, and the blob data that is returned by the query is from the
newest relevant blob, whether it be the tiddlywiki_file attachment
or the newest save from the saved_content_file attachment.

Some guidance taken from
https://www.geekytidbits.com/postgres-distinct-on/

Also added some test coverage that tries to demonstrate that it
actually works.