Nginx config so it knows about webpack packs also
They are not under the rails assets dir so they need to be listed
separately. This should let them get the correct cache headers and
be served in gzip format the same as the other assets.

Also, do an assets:clean when building to clean out old files, which
I didn't realize had been building up for many months.

Note: There's also assets:clobber which nukes the directory
entirely, but I think assets:clean is leaves a few versions in which
might be useful in future.