Support redirecting tspot sites
There are two ways to redirect:
1. To one of your own Tiddlyhost sites?
2. To an arbitrary url

I've decided to make the first method available for anyone, but the
second method is behing a feature flag so it can be enabled for
subscribed users in future.

The behavior for hub listed sites that redirect is questionable.
Currently they continue to appear in the hub even though they get
redirected. I guess it's okay.