Stop using deprecated replace_on_assign_to_many
It's not actually removed until Rails 7.1, but there is so much
deprecation output in the test suite that I want to fix it now.

The change IIUC is that using `some_attachment.attach(attachables)`
will append, but `update(some_attachment: attachables)` will replace
everything. Rather than use attach, I'm adding the existing blobs
explicitly. This is a getting quite messy, e.g. passing the site
object into the class method as a param is not very elegant to say
the least. It's motivating me to do some refactoring to clean it up,
but let's save it for later.