Cosmetic tweaks for the site create buttons
Also remove the extra create button at the top since I feel like it
doesn't really belong there next to the sort/filter controls, and
also because there's another create button close by in the main site
header.

The stylesheets and the styles in general are still quite a mess of
inconsistencies and duplications. Some day I'll fix that, but not
today I guess.