Don't join blobs table when creating admin lists
The "last update" sorting has been broken for a little bit. This
fixes the sorting, but without the join to the blobs table there are
some drawbacks, namely you can no longer sort by the blob size, and
also, because the updated date used for sorting is from the site
record, it is now touched when site settings change. Previously
only a site save would touch it.

Making the admin query work like the hub query, (with DISTINCT ON,
with_blobs_for_query and blob_coalesce) was how I originally planned
to fix this, but it turned out to be really difficult so I gave up
on making it work.

(Now I'm wondering if we can get away with removing the blobs table
join from the hub query to remove that whole complicated mess, so I
added a todo comment about that.)