Script to migrate all sites to new attach format
Now that I'm fairly confident that the new has_many_attached
association is working fine, it's time to migrate all sites and move
a step closer to removing the code for use the old association.

Interestingly there was one "broken" tspot site where the blob
didn't exist. Not sure how that happened, but I was able to restore
it from the old bucket.

Will check this in now as a reference, but probably no need to keep
it for long.

Note: I realize there might have been a clever way to move the blob
attachment without touching the s3 bucket at all, thereby saving
some aws traffic, but it would have been tricky to test and also
this method has the advantage of populating many missing
raw_byte_size, tw_version and tw_kind fields.