Fix jsMath downloads
There's something broken about the place where I used to fetch
jsMath from, though from the page at [1] it seems like it should
work.

Use sourceforge instead which is not a nice place but at least it
does still alloow curl downloads. I can't find the 1.2 version of
the fonts, but I'm hoping that the 1.3 version will work just as
well.

(At this point I think the GitHub workflow prod build might be
ready to use. Next steps is to push it and test it.)

Note: Downloading and unzipping this on every CI build seems like a
lot of overhead, especially considering I'm not sure how many people
care much about jsMath support for the few very old TiddlyWikis that
use it. Will stick with it for now, but maybe there's a more
efficient way to do it.

[1] https://www.math.union.edu/~dpvc/jsMath/download/old-versions.html