Fix encoding errors reading old Tiddlyspot files
Now that I have exception notifications via email, I've noticed the
some "invalid byte sequence in UTF-8" exceptions showing up for old
Tiddlyspot sites. Hopefully this fixes it.

Because of this problem I'll keep the connection to the old
dreamhost bucket for a while longer.

Required a rebuild of the base image to add the libicu-dev package
since it's a dependency of charlock_holmes.

References:
https://dev.to/bajena/solving-invalid-byte-sequence-in-utf-8-errors-in-ruby-1f27
https://stackoverflow.com/questions/31644103/ruby-converting-string-encoding-from-iso-8859-1-to-utf-8-not-working