Upgrade bootstrap and add dark mode toggle
Using the new-ish bootstrap 5.3.0-alpha1 for the dark mode support.
It works but it doesn't look good currently. Will be adding a lot of
CSS changes in future commits to make dark mode actually work.

Note: The bootstrap examples use local storage instead of a cookie,
and the switch the theme without performing a server fetch, but I'm
doing it the old-fashioned way with a server-side cookie set and
re-render. Will maybe revisit that later, and perhaps consider using
local storage for the grid view cookie as well.