More sane prod dependencies install on build
Instead of copying the gems and modules into the prod container, do
the installs inside the container while building.

This takes more time but I think this is a more sane way to do it,
and it avoids some weirdness such as doing the assets precompile in
the base image.

Notice that the credential secret workaround for compiling assets
without access to the real secret, (figured out initially for the
GitHub CI build), now needs to be used in the Dockerfile.

This change makes the CI build and the local build more similar, but
there are still some differences that I'd like eliminate in the
future, so there's likely some followup work needed.

Also includes an improvement to not install development
dependencies, both gems and node modules, in the prod container.

FYI I deployed a GitHub CI build to prod for the first time with
this.