Better deps installation in the prod dockerfile
In Rails 7.1 will have a default Dockerfile. I'm using some of the
methods used there via https://fly.io/ruby-dispatch/rails-on-docker/

The main change is related to installing the gems and modules
separately in their own layer. I was doing this before but it a more
complicated and weird way. This is is much better.

This is kind of a followup from recent commit "More sane prod
dependencies install on build".

Note: With this change, there's probably no good reason to continue
keeping the node_modules and the bundle dir in their non-default
locations. I'll aim to to move them back to their usual place, but
not in this commit.