Increase default number of site versions to keep
As per the commentary. Also it enables showing a preview version of
the site history page to free plan users as an incentive to consider
subscribing.

There was test that failed due to this change, but I removed it
since it was not really needed since site model test covers the same
code with better stubbing.