Rename old user plan to user_type
The user "plan" is schema that has been around forever. There are
two plans currently, "basic" and "superuser". Now that there are
subscriptions with subscription plans I want to change the
terminology to prevent confusion.

There was some extra work needed due to the fact that Plan.create!
is used in an old migration. I'm trying to solve that with a db
seed, (but worrying about the migration is probably not necessary,
see next commit).