Don't use db:migrate to set up a new environment
In the last commit there was some effort to make sure db:migrate
would continue to work in a brand new database, which is why I was
thinking about this.

Looking at [1] I think it's probably better to use db:setup, which
uses db/schema to setup the schema in one step, rather than
replaying every migration from the beginning.

[1] https://stackoverflow.com/questions/10301794/difference-between-rake-dbmigrate-dbreset-and-dbschemaload