Add delayed job background job handler
All this time we've been running without a proper background job
back end. Not sure exactly how it was working in prod, but for some
reason it stopped working recently, which means the prune old
versions job and the thumbnail generator haven't run for a while.

Not sure how to debug it so let's add a real active job back end.

Choosing delayed job over sidekiq or something more modern, firstly
since I'm a little familiar with it, and secondly because it can use
the main db as a backend rather than redis, so it's easier to
deploy.

The delayed job worker has its own container. Because it runs rails
it's nearly the same as the app container.

Note: The "seccomp:unconfined" is probably only needed in the jobs
container now, since it's for using chrome to generate thumbnails,
and that happens in a background job, but the tests fail without it
in the app container too, since the tests try to generate a thumbnail
immediately in the foreground.

Todo maybe: Consider tweaking the job priorities and whether
multiple queues would be useful.