Improve subscription payment email text
Convert the views to haml and make the content less generic.

Also add some basic test coverage similar to the devise email
coverage, which required adding some pay related fixtures.

Some of these emails are not used, but I updated them anyway for
consistency and in case they get used in the future.