Don't show all versions in save history preview
Actually do show few more than the limit but with a fade-out effect.
It would be seen right after you cancelled your subscription and
lost the save history feature but you haven't yet saved and
triggered the prune of old versions.

(This is kinda cheesy but actually there are some serious UX
considerations around the possibility of losing those old versions
when you save with the history feature disabled though.)