Mount home .cache dir outside dev container
I was debugging why puppeteer didn't work after a fresh build of the
base image, and it's because it didn't have its chrome installed in
~/.cache/puppeteer.

Doing a yarn install doesn't actually fix it, since yarn thinks
everything is installed correctly already, which isn't true if the
content of ~/.cache/puppeteer is gone.

Mounting the ~/.cache outside the container means its content will
still be there after a base image rebuild.