Workaround for unexpected error after subscription
The error appeared for a new subscriber. I suspect it was when they
returned to Tiddlyhost immediately after subscribing. One theory is
that maybe there are multiple non-atomic webhooks after a
subscription event and the error happened somewhere between the
first and the last, but that's just a theory though, some
investigation is needed.

As per the commentary, the workaround is not a good one, and could
cause weirdness when/if there is more than one stripe product. But
for now I think it's better than showing an "Internal server error"
right after a suscription.