Fix external core script tag finder
Fix a bug that has been around since the TW 5.2.6 upgrade causing
the inject_external_core_url_prefix method to silently do nothing
for 5.2.6 and 5.2.7 sites, which meant the external core urls were
relative instead of being translated to tiddlyhost.com/corefile.js.

Also add some test coverage that would have caught this problem, and
do some extra work to future-proof TiddlyWiki version comparisons.

Note: It would have been possible to use "//script" and find the
script tag wherever it is, but since the doc is large, I figure
searching all of it would be a bad idea, so that's why we're keeping
the explicit path.