Make it easier to move var lib docker locally
After a dnf update, docker's config reverts back to its default, so
it goes back to using /var/lib/docker instead of my custom location
which is /home/docker.

The ansible to fix it in prod is the same thing I want to do
locally. This arguably doesn't belong in this repo, but it does make
my life easier.