Skip uglify when generating TW 5.3.0 external js
I think tw5-uglify plugin is buggy for the new TiddlyWiki 5.3.0. It
appears there are some encoded '<' chars that uglify converts back
to plain '<' and somehow it breaks things.

This does increase the size substantially, from around 1.3k to
around 2.1k. It is likely that most browsers are fetching it gzipped
anyhow.

I'll aim to report a bug with tw5-uglify and see if this can be
fixed.