Avoid email rendering error on failed payment
I haven't tested this thoroughly but I believe that the web hook for
a failed payment triggers an email which throws an error because the
pay_charge param is not present.

This is a band-aid fix, but hopefully it will help avoid the error
being thrown. (I saw a lot of these errors on a failed payment due
to delayed job retries.)