Update manifest.js to fix sprockets error
The error is this:
  Error: Asset 'application.css' was not declared to be precompiled
  in production

Not sure what the root cause is. In
https://github.com/rails/sprockets-rails/issues/458 it suggests it
might be related to a sprockets-rails update, but I don't see that
happening here.

Since all the stylesheets are imported via application.scss, I think
the change makes sense.

Saw a similar issue for the bootstrap icons, so fix it in the same
way.