Option to show both internal & external core tw5
Add a special filter option for the hub to include both the tw5 and
tw5x kinds. I considered use this only and hiding the more specific
kind filters, but I think it is useful when choosing a template to
be able to select one or the other, so I decide to show all three
even though it does make it a bit cluttered.