Improve handling of recaptcha failure
The expected workflow is that when there is a recapture failure the
user gets a chance to try again. You can reproduce this by waiting
two minutes between loading the form and submitting the form, since
that's long enough to get a "timeout-or-duplicate" error code from
recaptcha.

This patch adds a workaround so that the previously entered name and
username fields are preserved when that happens. The user still
needs to reenter the two password fields, but I think that's fine.