Pin a base image digest in the prod Docker file
Previously, whenever I rebuild the base image, there's no sign of it
in the commit history. This makes it so there's a commit showing
that it happened and also making the prod build a little more
reproducible.

Include moving some bash out of the Makefile into a script to make
it easier to read.

(Note: I was surprised to learn there's no actual image digest until
the image gets pushed, so that's why I'm pushing the base image to
docker now. It seems a little strange, but I guess it should be
okay.)