classic: Don't always force visible edit buttons
If you're viewing a TiddlyWiki Classic site isn't your own, or is
your own but you're not logged in, previously the edit buttons would
be shown regardless of the value of the chkHttpReadOnly "advanced
options" cookie.

Now, the chkHttpReadOnly value will be respected, unless the site
owner is currently logged in.

The way it works is to inject a shadow tiddler with content set to
either 'yes' or 'no', then read the value of that tiddler in the
upload plugin. This is similar to the TiddlyWiki5 method, except in
TiddlyWiki5 the tiddler used is '$:/status/IsLoggedIn'.

Closes issue #326.