Fix encoding errors for some old TiddlySpot sites
Some old sites are being detected by CharlockHolms as IBM424_rtl.
This may be accurate, but I doubt it.

Either way it throws errors if we try to convert it since ruby
doesn't support that encoding, so let's not be so eager to convert
things.

More about ruby encoding support here:
https://ruby-doc.org/core-2.5.1/Encoding.html