Modal dialog editing for the saved version labels
- A general purpose-ish wrapper for bootstrap modals
- A neat way to open them and set their body content dynamically
- Controller actions to serve the modal form and save the label
  update

An integration test would perhaps be better than the controller test
included here, but I think it's probably good enough.

Note that I'm not doing the post-change refresh dynamically, so
there's a whole page reload after the change. This means the code is
a little simpler and less javascript is needed, but the downside is
the UI is less slick and there are more server round trips than
there could be. Maybe we can revisit that later, but I am enjoying
the fact that this patch has only two lines of javascript.