Keep labeled versions when pruning saved versions
This is kind of the main purpose of the labels. The idea is that you
can label an important revision, e.g. something like "first draft"
and have it kept indefinitely.

The oldest labeled revision could still be removed if every single
revision has a label, but I'm not expecting people would have 100
labeled revisions.