Provide a favicon when viewing saved TW versions
TiddlyWiki files for some reason include a local ref to a favicon,
i.e., <link ... href="favicon.ico">. The routing for viewing a saved
version from the save history page was causing this to incorrectly
match the blob id route and produce a 500 server error since
favicon.ico is not a valid blob id.

Fix it by adding a dedicated route for the favicon request.

Adjusting the link tag on the fly might provide a more elegant
solution, but let's keep it simple for now.