Initial support for per-site storage service
The motivation is so I can be more flexible about the storage
service used, experiment with different providers, and perhaps
eventually migrate to a less expensive provider than the one used
currently.

Note: Another way to do this would be to specify the service as a
proc in the has_many_attached association, see [1], but I decided
not to do that. This approach seems simpler and should work just
as well. Also my first attempt to use a proc in the association hit
some problems that I didn't feel like debugging.

[1] https://github.com/rails/rails/commit/8bf1353cae3736045cc8343a7bca5d1855497214