Adjust read-only behavior for TiddlyWiki classic
This rolls back the behavior change from #328, but provides a way
to bring it back if you prefer it, by setting up a window.bidix
object in another tiddler tagged with "systemConfig".

As mentioned in the comments, the motivation for doing this is that
hiding the "upload to tiddlyhost" button causes usability problems
for TW classic users.