Move logging middleware to app/middleware
I'm not confident I understand all the implications of eager loading
versus autoloading, but this seems to be okay. (It might be that I
should be adding the extra paths to the auto load list instead of
the eager load list. It might also be that I don't really need the
middleware class to be eager loaded.)

The reason to do this now is because I'm adding another middleware
class in the next commit and I want to be tidy and put them in a
specific place rather with the rails/lib miscellany.

Also:
- Add a .to_s to the paths. I noticed all the other paths in
  ActiveSupport::Dependencies.autoload_paths are strings, so let's
  me consistent.
- Add a config.autoload_lib(ignore: ..) based on the docs at
  https://guides.rubyonrails.org/autoloading_and_reloading_constants.html#config-autoload-lib-ignore