Fix bug preventing account updates from working
I'm pretty sure this used to work, but maybe something changed
related to how rails handles the before actions work. What I think I
was observing was that the before action would not run at all in the
registrations controller. I'm assuming that's related to it being
added twice with different conditions.