Add json endpoint for listing all your sites
The initial motivation is to be able to use it to script a "backup
all sites" shell script, similar to examples/thost-downloader.

Note that the individual site end point was already working, which
was something of a suprise. Appologies if you were using that and I
just broke your stuff!