Download a pre-release core js when building
For the prerelease empty, a fresh copy is fetched regularly, (see
bin/fetch-prerelease-empty.j2 in the ansible templates directory)
but we're not doing that for the prerelease core js file currently,
so it's likely going to drift from the html file.

Maybe in future we'll do that, but in reality I'm not expecting many
users will need or use this file. That might change if an "external
core pre-release" empty option is made available.