Remove docker compose --progress plain flag
I can't recall exactly why I wanted it, perhaps for debugging. Let's
remove it now, but make it easy to use it if you want.

I might have wanted it for the (currently unused) GitHub action
build, but the default "auto" option should detect that and use
plain there anyhow.

Also: There was a warning to say that should be a global compose
flag, so it was moved before the build command.