More easily customizable nginx config
Previously I was leaving the /etc/nginx/nginx.conf in the nginx base
image untouched and putting the Tiddlyhost specific config in
/etc/nginx/conf.d only.

It this patch I've started explictly adding the full nginx.conf file
and moved the Tiddlyhost specific config into it. able to customize
the log format, which is not possible afaict unless you can modify
the original access_log command.

Note: I was able to add additional logging with a custom format in
conf.d, but removing the originally defined logging is (aiui) not
possible, so I was seeing two lines logged for every request, one
with the original format and one with the custom format. In any
case, it's probably better this way. It will make it easier to
customize the nginx config in future.