Rework node_modules volume mount, remove symlink
Adjust the development volume mount and remove the need for a
symlink. Motivated by the fact that symlinks don't work well on
Windows systems.

I was using a symlink to avoid having the node_modules volume
mounted inside the rails volume, but actually that works fine.

I also moved the directory inside docker just for consistency with
the other similar volume mounts.