Add advanced setting to skip etag check
Sometimes people get a "File changed on server" error that they
can't get rid of.

I'm not sure what the root cause is, but it could be that the
browser is unexpectedly using a cached version of the site instead
of the latest version, and hence has a stale etag, or it could be a
weird bug somewhere related to the etag handling.

Whatever the case, the goal here is to provide a workaround for
users in the stressful situation of being unable to save their
changes.

Do that by adding an option to skip the etag comparison and just
save anyway. It's not a convenient solution, but better than
nothing I think.

(Also remove an extraneous full-stop in some form text for
consistency.)

See also:
https://talk.tiddlywiki.org/t/cannot-save-to-tiddlyhost/11371