Fix site count inconsistencies when searching
If you do a search on your sites list, the value of @sites.count
would be the number of sites that matched the search, rather than
the total number of sites.

This caused subtle UI misbehaviors, e.g. the message "Showing X of Y
sites total".

The fix is a bit clumsy, but I want to avoid reimplementing the
searching logic outside of the hub query SQL, especially because it
would also have to handle merging the two different site models,
which makes it quite difficult.