Show thumbnail of clone from site for new sites
Thumbnail generation jobs are queued with a two minute delay,
so when a brand new site is created it has no thumbnail.

But, for cloned sites, the thumbnail of the site that was cloned is
actually the perfect thumbnail for the new site, so let's just show
it until the real thumbnail is created.