Stop using non-standard docker data root
The old server had limited space on the root volume, so using
/var/lib/docker for docker images would fill it up. To deal with
that I configured docker to use /home/docker instead of
/var/lib/docker, since the home volume had plenty of space.

Now that we no longer have a small root partition, this workaround
is not needed, so let's remove it.

Also, it avoids needing to re-run these specific ansible tasks after
upgrading docker, and should avoid a major problem I've seen
recently when the /home volume is not available when the docker
service is starting up after a reboot.