Add an allow list for recaptcha checking
After googling the "browser-error" error code, which is what I'm
seeing in the logs for this particular user, it seems like there are
other cases of it being diagnosed as a unexpected failure. Some
mention it could be ad-blockers or Pi-holes. In any case this should
be fine to allow one specific user having problems.