More syncing of rails defaults with app:update
Update some rails boot and environment setup scripts to be closer to
the rails defaults.

Not confident I know all the details about the changes, but I'm
asking a popular LLM for explanations, and so far it seems like it
should be fine.

There are still more changes to come, but I'm once again trying to
cut this into smaller chunks and merge the low risk parts first.

A bunch of this of this is just comment changes which of coures have
no significant impact. Also some development and test environment
changes should be low risk for breaking prod.