Add task to precompile bootstrap email sass cache
Because of this issue[1], the first time I run the tests suite after
a `make cleanup` (and hence after a `make deps-update`) it spits out
a huge amount deprecation warnings. Because of all the warnings,
it's hard to see the actual test suite results, so what I've been
doing is running the test suite twice. The second time it runs, the
styleheets are already cached, so the output is clean and tidy.

This change address the annoyance by allowing the noisy sass
precompile to be done first before running the tests.

[1] https://github.com/bootstrap-email/bootstrap-email/issues/277