Show user email confirmed timestamp in admin list
Make a user's confirmation status visible to the admin user. Show
the time between created and confirmed since it's more interesting.

I've been assuming people with zero logins were unable to confirm
their email, but it's not always the case. Actually there are a
significant number of users who confirmed their email address but
never logged in, which is surprising.

(I guess it would be possible to sort by the created/confirmed
timespan rather than the raw timestamp, but I'm not going to
bother.)